HARMONY-380: Fix cloud access endpoints, switch to better mock for AWS calls, and cleanup README.
HARMONY-380: We should keep Patrick around, but this makes the AWS mock work on the first load of the code.

HARMONY-380: Remove another import * as aws

HARMONY-380: Test not setting the SDK instance in the tests.

HARMONY-380: Give up trying to stub at the AWS level. Stub the SecureTokenService wrapper around AWS.

HARMONY-380: Cleanup response that no longer needs all of the fields.